requirements.txt
addict==2.4.0
affine==2.4.0
ansi2html==1.8.0
appdirs==1.4.4
asciitree==0.3.3
asttokens==2.2.1
attrs==23.1.0
backcall==0.2.0
boto3==1.28.15
botocore==1.31.15
certifi==2023.7.22
cftime==1.6.2
charset-normalizer==3.2.0
click==8.1.6
click-plugins==1.1.1
cligj==0.7.2
cloudpickle==2.2.1
comm==0.1.3
ConfigArgParse==1.7
configobj==5.0.8
contourpy==1.1.0
cycler==0.11.0
dash==2.11.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dask==2023.7.1
debugpy==1.6.7
decorator==5.1.1
docutils==0.20.1
donfig==0.8.1
entrypoints==0.4
eumdac==2.0.1
executing==1.2.0
fasteners==0.18
fastjsonschema==2.18.0
Flask==2.2.5
fonttools==4.41.1
fsspec==2023.6.0
h5py==3.9.0
idna==3.4
importlib-metadata==6.8.0
ipykernel==6.25.0
ipython==8.14.0
ipywidgets==8.0.7
itsdangerous==2.1.2
jedi==0.19.0
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.1
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyterlab-widgets==3.0.8
kiwisolver==1.4.4
locket==1.0.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
mdurl==0.1.2
memory-profiler==0.61.0
meshio==5.3.4
nbformat==5.7.0
nest-asyncio==1.5.7
netCDF4==1.6.4
numcodecs==0.11.0
numpy==1.25.1
open3d==0.17.0
opencv-python==4.8.0.74
packaging==23.1
pandas==2.0.3
parso==0.8.3
partd==1.4.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==3.9.1
plotly==5.15.0
pooch==1.7.0
prompt-toolkit==3.0.39
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
PyGLM==2.7.0
Pygments==2.15.1
pykdtree==1.3.7.post0
PyOpenGL==3.1.7
PyOpenGL-accelerate==3.1.7
pyorbital==1.8.0
pyparsing==3.0.9
pyproj==3.6.0
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.2
pyquaternion==0.9.9
pyresample==1.27.1
pyspectral==0.12.3
python-dateutil==2.8.2
python-geotiepoints==1.6.0
pytz==2023.3
pyvista==0.41.1
PyYAML==6.0.1
pyzmq==25.1.0
rasterio==1.3.8
referencing==0.30.0
requests==2.31.0
retrying==1.3.4
rich==13.5.0
rpds-py==0.9.2
s3transfer==0.6.1
satpy==0.43.0
scikit-learn==1.3.0
scipy==1.11.1
scooby==0.7.2
shapely==2.0.1
six==1.16.0
snuggs==1.4.7
stack-data==0.6.2
tenacity==8.2.2
threadpoolctl==3.2.0
toolz==0.12.0
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
trollimage==1.20.1
trollsift==0.5.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
vtk==9.2.6
wcwidth==0.2.6
Werkzeug==2.2.3
widgetsnbextension==4.0.8
wxPython==4.2.1
xarray==2023.7.0
zarr==2.16.0
zipp==3.16.2
